requirements.txt
# Environment and Configuration
python-dotenv

# RSS Parsing
feedparser

# Web Scraping
scrapy

# WordPress API and Network Requests
requests
